# leargon-backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Léargon API
  description: 'REST API for Léargon data and process governance platform.


    ## Authentication

    The API uses JWT (JSON Web Token) based authentication. After logging in or signing up, you will receive an access token
    that must be included in the `Authorization` header as a Bearer token for protected endpoints.


    ## Authorization Roles

    - **ROLE_USER**: Standard user role, assigned to all users by default

    - **ROLE_ADMIN**: Administrative role with full user management capabilities


    ## Fallback Administration Protection

    The system includes a fallback admin user that cannot be modified or deleted through the API. Any attempt to modify this
    user will result in a 403 Forbidden response.

    '
  version: 1.0.0
  contact:
    name: Léargon
servers:
- url: http://localhost:8081
  description: Development server
- url: http://localhost:8080
  description: Local container server
...
paths:
  /administration/users:
    get:
      tags:
      - Administration
      summary: Get all users
      description: Returns a list of all users in the system. Requires ROLE_ADMIN.
      operationId: getAllUsers
      security:
      - bearerAuth: []
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions (requires ROLE_ADMIN)
    post:
      tags:
      - Administration
      summary: Create a new user
      description: Creates a new local user account. Works even when Azure authentication is configured. Requires ROLE_ADMIN.
      operationId: createUser
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions (requires ROLE_ADMIN)
        '409':
          description: Email or username already exists
  /administration/users/{id}:
    get:
      tags:
      - Administration
      summary: Get user by ID
      description: Returns detailed information about a specific user. Requires ROLE_ADMIN.
      operationId: getUserById
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: User ID
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions (requires ROLE_ADMIN)
    put:
      tags:
      - Administration
      summary: Update user details
      description: Updates user information including name, email, username, status, and roles. Cannot modify fallback admin
        user. Requires ROLE_ADMIN.
      operationId: updateUser
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: User ID
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          description: Cannot modify fallback admin user
        '404':
          description: User not found
        '409':
          description: Email or username already exists
        '401':
          description: Not authenticated
    delete:
      tags:
      - Administration
      summary: Hard-delete user
      description: Permanently deletes a user account. Only allowed if the user has no data ownership, process ownership, or organisational unit lead assignments. Cannot delete fallback admin user. Requires ROLE_ADMIN.
      operationId: deleteUser
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: User ID
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: User permanently deleted
        '403':
          description: Cannot delete user — either fallback admin or user still has ownership/lead assignments
        '404':
          description: User not found
        '401':
          description: Not authenticated
  /administration/users/{id}/disable:
    post:
      tags:
      - Administration
      summary: Disable user account
      description: Disables a user account, preventing login. Cannot disable fallback admin. Requires ROLE_ADMIN.
      operationId: disableUser
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: User ID
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Account disabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          description: Cannot disable fallback admin
        '404':
          description: User not found
        '401':
          description: Not authenticated
  /administration/users/{id}/enable:
    post:
      tags:
      - Administration
      summary: Enable user account
      description: Enables a disabled user account, allowing login. Requires ROLE_ADMIN.
      operationId: enableUser
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: User ID
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Account enabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
        '401':
          description: Not authenticated
  /administration/field-configurations:
    get:
      tags:
      - Administration
      summary: Get all field configurations
      description: Returns the list of mandatory field configurations per entity type. Requires ROLE_ADMIN.
      operationId: getFieldConfigurations
      security:
      - bearerAuth: []
      responses:
        '200':
          description: List of field configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldConfigurationEntry'
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions (requires ROLE_ADMIN)
    put:
      tags:
      - Administration
      summary: Replace all field configurations
      description: Replaces the complete list of mandatory field configurations. Requires ROLE_ADMIN.
      operationId: replaceFieldConfigurations
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/FieldConfigurationEntry'
      responses:
        '200':
          description: Updated list of field configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldConfigurationEntry'
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions (requires ROLE_ADMIN)
  /administration/users/{id}/password:
    put:
      tags:
      - Administration
      summary: Change user password (admin)
      description: Allows an administration to reset a user's password without knowing the current password. Cannot change
        fallback admin password. Requires ROLE_ADMIN.
      operationId: administrationChangePassword
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: User ID
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdministrationChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
        '403':
          description: Cannot change fallback admin password
        '404':
          description: User not found
        '400':
          description: Invalid password (must be at least 8 characters)
        '401':
          description: Not authenticated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /authentication/login or /authentication/signup
  schemas:
...
    AdministrationChangePasswordRequest:
      type: object
      required:
      - newPassword
      properties:
        newPassword:
          type: string
          format: password
          minLength: 8
          description: New password (minimum 8 characters)
...
    SignupRequest:
      type: object
      required:
      - email
      - username
      - password
      - firstName
      - lastName
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address (must be unique)
          example: user@example.com
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Username (must be unique)
          example: johndoe
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 255
          description: Password (minimum 8 characters)
          example: password123
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          description: User's first name
          example: John
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          description: User's last name
          example: Doe
...
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: New email address (must be unique if changed)
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: New username (must be unique if changed)
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          description: New first name
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          description: New last name
        preferredLanguage:
          type: string
          description: Locale code (e.g., 'en', 'de')
          example: en
        enabled:
          type: boolean
          description: Account enabled status
        roles:
          type: array
          description: User roles
          items:
            type: string
            enum:
            - ROLE_USER
            - ROLE_ADMIN
          example:
          - ROLE_USER
          - ROLE_ADMIN
    UserResponse:
      type: object
      required:
      - id
      - email
      - username
      - firstName
      - lastName
      - preferredLanguage
      - enabled
      - isFallbackAdministrator
      - setupCompleted
      - roles
      - createdAt
      - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: User ID
          example: 1
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        username:
          type: string
          description: Username
          example: johndoe
        firstName:
          type: string
          description: User's first name
          example: John
        lastName:
          type: string
          description: User's last name
          example: Doe
        preferredLanguage:
          type: string
          description: Locale code (e.g., 'en', 'de')
          example: en
        enabled:
          type: boolean
          description: Account enabled status
          example: true
        isFallbackAdministrator:
          type: boolean
          description: Whether this is the protected fallback administrator user
          example: false
        setupCompleted:
          type: boolean
          description: Whether this user has completed the first-time setup wizard
          example: false
        roles:
          type: array
          description: User roles
          items:
            type: string
          example:
          - ROLE_USER
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: '2024-01-15T10:30:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: '2024-01-15T10:30:00Z'
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          description: Last login timestamp
          example: '2024-01-15T14:20:00Z'
        authProvider:
          type: string
          nullable: true
          description: Authentication provider (e.g., AZURE for Azure Entra ID users)
          example: null
...
    FieldConfigurationEntry:
      type: object
      required:
      - entityType
      - fieldName
      properties:
        entityType:
          type: string
          description: Entity type (BUSINESS_ENTITY, BUSINESS_DOMAIN, BUSINESS_PROCESS, ORGANISATIONAL_UNIT)
          example: BUSINESS_ENTITY
        fieldName:
          type: string
          description: Field name that is mandatory (e.g. retentionPeriod, businessDomain, descriptions)
          example: retentionPeriod
...