# leargon-frontend/src/api/generated/model/fieldConfigurationEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Léargon API
 * REST API for Léargon data and process governance platform.

## Authentication
The API uses JWT (JSON Web Token) based authentication. After logging in or signing up, you will receive an access token that must be included in the `Authorization` header as a Bearer token for protected endpoints.

## Authorization Roles
- **ROLE_USER**: Standard user role, assigned to all users by default
- **ROLE_ADMIN**: Administrative role with full user management capabilities

## Fallback Administration Protection
The system includes a fallback admin user that cannot be modified or deleted through the API. Any attempt to modify this user will result in a 403 Forbidden response.

 * OpenAPI spec version: 1.0.0
 */

export interface FieldConfigurationEntry {
  /** Entity type (BUSINESS_ENTITY, BUSINESS_DOMAIN, BUSINESS_PROCESS, ORGANISATIONAL_UNIT) */
  entityType: string;
  /** Field name that is mandatory (e.g. retentionPeriod, businessDomain, descriptions) */
  fieldName: string;
}